double cast to a cieling functiont to avoid segfault

+++ my_malloc.c
@@ -318,7 +318,7 @@ void* myallocate(int bytes, char * file, int line, int req){
 			// We need to add more pages
 			// If the prev was free, increase the size of prev
 			if (((SegMetadata *)prev)->used == BLOCK_FREE) {
-				reqPages = ceil((bytes - ((SegMetadata *)prev)->size)/PAGESIZE);
+				reqPages = ceil((double)(bytes - ((SegMetadata *)prev)->size)/PAGESIZE);
 				((SegMetadata *)prev)->size += reqPages * PAGESIZE;
 				ptr = prev;
 			}
